fix range handling
the early range handling (needs syntax and merge on same paths) had an
error detecting the range segment due being inherently flawed in the
first implementation. it was working for the job, but leaded to errors on
non-existing paths.

fix is to properly split path and range from each other if applicable by
better validation and a safer string split (from behind).