fix pascii
the pascii (print ascii) aims to print in the printable set of characters
by converting C0, C1 and high into hex sequences like \x##.

However the output can not be translated back stable as the backslash
"\" (\x5C) was not translated previously when could have also being in
the form representing \x00 sequences (no way back).

Fix is to convert the backslash ("\") as well to the hex-sequence when it
is followed by characters that would make it a matching \x[0-9A-F]{2}
sequence.