less smoke on git archive
when fiddling w/ git attributes configuration in light of creating
packages w/ `$ git archive ...` it's nice to see whether or not uncommitted
changes create the wished result or not.

therefore the smoke test for a fine package is done before the real test
with the same test but taking the working tree git attributes into account.