better support for git core.quotePath
in c90ce5b support for git core.quotePath handling has been introduced.

albeit it is functionality wise working, it is applied too late.

for example a command which contains an operation on a file earlier, was
spilling errors:

    $ git ls-files | ./pcre.php --only '/this file is part of pcre\.php/'

the --only path filter was still operating on non-normalized paths.

improvement is to move the path mapping towards the very beginning of path
processing and map files' paths as early as possible to come towards a path
normalization.

ref: c90ce5bb852aef9295110ee2b0fbf81d1e014f5b