better support for null bytes in paths
a path on it's own is considered invalid if it contains a NUL byte.

this went unchecked in larger parts. attempt here is to add a path
validation function and to start using it.

for example with a wrapper for php's file() function which otherwise would
fatal error on such paths.

also made the fuzzy_quote_path_normalizer() a function and improved it
so that it does not fatal error but also not unquote for invalid paths as
those invalid paths are more a clear sign for NUL byte separated paths and
not newline separated paths. so an invalid path can never be a quote path.