work with null byte separated paths
like other utilities, git-ls-files supports terminating a path with a null
byte (-z) comparable to find with -print0.

previously pcre.php did not specifically handle such input expecting by
default that paths are terminated by a newline.

when null bytes are involved for paths input, it can be safely expected
that the null byte is a path terminator as valid path names can not
contain null bytes.

change is to do a buffered read on paths inputs (4096 bytes) to detect
whether or not null bytes are used used as path terminators and if not,
expect (like before) that newlines are used as path terminators.