fix -m, --multiple and add support for --show-matches
previously the -m, --multiple was only supported for search and replace,
but not for --show-matches.

additionally the match for decimals has been fixed. previously aggregation
was done on the match itself as array keys. this converts into an integer
if the match represents a number (int or float). fix is to prefix the
match by an underscore character so that this unwanted conversion does
not kick in.